tech: moved some files, added animations scss file.
