+++ public/styles/_global.scss
@@ -1,4 +1,6 @@
 @use './colors' as *;
+@use './font' as *;
+@use './animations' as *;
 
 @forward './body.scss';
