+++ src/components/global/index.ts
@@ -1,2 +1,2 @@
-export {default as TheForm} from './TheForm.vue'
-export {default as TheMenu} from './TheMenu.vue'
+export {default as TheForm} from './reusable/TheForm.vue'
+export {default as TheMenu} from './used-once/TheMenu.vue'
